+++ MapboxSearch/sample/src/main/java/com/mapbox/search/sample/api/JapanSearchKotlinExampleActivity.kt
@@ -5,13 +5,13 @@ import android.os.Bundle
 import android.util.Log
 import com.mapbox.search.ApiType
 import com.mapbox.search.Country
-import com.mapbox.search.Language
 import com.mapbox.search.ResponseInfo
 import com.mapbox.search.SearchEngine
 import com.mapbox.search.SearchEngineSettings
 import com.mapbox.search.SearchOptions
 import com.mapbox.search.SearchSelectionCallback
 import com.mapbox.search.common.AsyncOperationTask
+import com.mapbox.search.common.IsoLanguage
 import com.mapbox.search.result.SearchResult
 import com.mapbox.search.result.SearchSuggestion
 import com.mapbox.search.sample.R
@@ -65,7 +65,7 @@ class JapanSearchKotlinExampleActivity : Activity() {
             "東京",
             SearchOptions(
                 countries = listOf(Country.JAPAN),
-                languages = listOf(Language.JAPANESE),
+                languages = listOf(IsoLanguage.JAPANESE),
             ),
             searchCallback
         )