+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/ReverseGeocodingSearchIntegrationTest.kt
@@ -10,6 +10,7 @@ import com.mapbox.search.base.utils.orientation.ScreenOrientation
 import com.mapbox.search.base.utils.orientation.ScreenOrientationProvider
 import com.mapbox.search.common.AsyncOperationTask
 import com.mapbox.search.common.FixedPointLocationEngine
+import com.mapbox.search.common.IsoLanguage
 import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.common.SearchRequestException
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
@@ -105,7 +106,7 @@ internal class ReverseGeocodingSearchIntegrationTest : BaseTest() {
         val options = ReverseGeoOptions(
             center = TEST_POINT,
             countries = listOf(Country.UNITED_STATES, Country.BELARUS),
-            languages = listOf(Language.ENGLISH),
+            languages = listOf(IsoLanguage.ENGLISH),
             limit = 5,
             types = listOf(QueryType.ADDRESS, QueryType.POI)
         )
@@ -119,7 +120,7 @@ internal class ReverseGeocodingSearchIntegrationTest : BaseTest() {
         val url = request.requestUrl!!
         assertEqualsIgnoreCase("//search/v1/reverse/${formatPoints(TEST_POINT)}", url.encodedPath)
         assertEquals(TEST_ACCESS_TOKEN, url.queryParameter("access_token"))
-        assertEquals(Language.ENGLISH.code, url.queryParameter("language"))
+        assertEquals(IsoLanguage.ENGLISH.code, url.queryParameter("language"))
         assertEquals(options.limit.toString(), url.queryParameter("limit"))
         assertEquals(
             options.types?.joinToString(separator = ",") { it.name.lowercase(Locale.getDefault()) },
@@ -220,7 +221,7 @@ internal class ReverseGeocodingSearchIntegrationTest : BaseTest() {
                 query = formatPoints(TEST_POINT),
                 endpoint = "reverse",
                 options = SearchOptions(
-                    languages = listOf(Language(Locale.getDefault().language)),
+                    languages = listOf(IsoLanguage(Locale.getDefault().language)),
                     proximity = TEST_POINT,
                     origin = TEST_POINT
                 ),