+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/SearchEngineIntegrationTest.kt
@@ -13,6 +13,7 @@ import com.mapbox.search.base.utils.orientation.ScreenOrientation
 import com.mapbox.search.base.utils.orientation.ScreenOrientationProvider
 import com.mapbox.search.common.AsyncOperationTask
 import com.mapbox.search.common.FixedPointLocationEngine
+import com.mapbox.search.common.IsoLanguage
 import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.common.SearchRequestException
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
@@ -124,7 +125,7 @@ internal class SearchEngineIntegrationTest : BaseTest() {
             boundingBox = BoundingBox.fromPoints(Point.fromLngLat(10.0, 15.0), Point.fromLngLat(30.0, 50.0)),
             countries = listOf(Country.UNITED_STATES, Country.BELARUS),
             fuzzyMatch = true,
-            languages = listOf(Language.ENGLISH),
+            languages = listOf(IsoLanguage.ENGLISH),
             limit = 5,
             types = listOf(QueryType.COUNTRY, QueryType.LOCALITY, QueryType.ADDRESS),
             origin = Point.fromLngLat(50.123, 70.123),
@@ -150,7 +151,7 @@ internal class SearchEngineIntegrationTest : BaseTest() {
             url.queryParameter("bbox")
         )
         assertEquals(options.countries?.joinToString(separator = ",") { it.code }, url.queryParameter("country"))
-        assertEquals(Language.ENGLISH.code, url.queryParameter("language"))
+        assertEquals(IsoLanguage.ENGLISH.code, url.queryParameter("language"))
         assertEquals(options.limit.toString(), url.queryParameter("limit"))
         assertEquals(
             options.types?.joinToString(separator = ",") { it.name.lowercase(Locale.getDefault()) },