+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/analytics/AnalyticsServiceImplTest.kt
@@ -6,7 +6,6 @@ import com.mapbox.search.ApiType
 import com.mapbox.search.BaseTest
 import com.mapbox.search.Country
 import com.mapbox.search.EtaType
-import com.mapbox.search.Language
 import com.mapbox.search.MapboxSearchSdk
 import com.mapbox.search.QueryType
 import com.mapbox.search.RouteOptions
@@ -20,6 +19,7 @@ import com.mapbox.search.base.utils.KeyboardLocaleProvider
 import com.mapbox.search.base.utils.orientation.ScreenOrientation
 import com.mapbox.search.base.utils.orientation.ScreenOrientationProvider
 import com.mapbox.search.common.FixedPointLocationEngine
+import com.mapbox.search.common.IsoLanguage
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
 import com.mapbox.search.tests_support.BlockingCompletionCallback
 import com.mapbox.search.tests_support.BlockingSearchSelectionCallback
@@ -80,7 +80,7 @@ internal class AnalyticsServiceImplTest : BaseTest() {
             navigationOptions = TEST_NAV_OPTIONS,
             boundingBox = BoundingBox.fromLngLats(0.0, 0.0, 90.0, 45.0),
             countries = listOf(Country.FRANCE, Country.GERMANY),
-            languages = listOf(Language.ENGLISH),
+            languages = listOf(IsoLanguage.ENGLISH),
             limit = 6,
             types = listOf(QueryType.POI, QueryType.ADDRESS, QueryType.POSTCODE),
             routeOptions = RouteOptions(
@@ -125,7 +125,7 @@ internal class AnalyticsServiceImplTest : BaseTest() {
             navigationOptions = TEST_NAV_OPTIONS,
             boundingBox = BoundingBox.fromLngLats(0.0, 0.0, 90.0, 45.0),
             countries = listOf(Country.FRANCE, Country.GERMANY),
-            languages = listOf(Language.ENGLISH),
+            languages = listOf(IsoLanguage.ENGLISH),
             limit = 6,
             types = listOf(QueryType.POI, QueryType.ADDRESS, QueryType.POSTCODE),
             routeOptions = RouteOptions(