+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/CategorySearchOptions.kt
@@ -7,6 +7,7 @@ import com.mapbox.search.Reserved.Flags.SBS
 import com.mapbox.search.base.core.CoreSearchOptions
 import com.mapbox.search.base.utils.extension.mapToCore
 import com.mapbox.search.base.utils.extension.safeCompareTo
+import com.mapbox.search.common.IsoLanguage
 import kotlinx.parcelize.Parcelize
 
 /**
@@ -44,7 +45,7 @@ public class CategorySearchOptions @JvmOverloads public constructor(
      *
      * Note: Geocoding API supports a few languages, Single Box Search – only one.
      */
-    public val languages: List<Language>? = defaultSearchOptionsLanguage(),
+    public val languages: List<IsoLanguage>? = defaultSearchOptionsLanguage(),
 
     /**
      * Specify the maximum number of results to return, including results from [com.mapbox.search.record.IndexableDataProvider].
@@ -122,7 +123,7 @@ public class CategorySearchOptions @JvmOverloads public constructor(
         boundingBox: BoundingBox? = this.boundingBox,
         countries: List<Country>? = this.countries,
         fuzzyMatch: Boolean? = this.fuzzyMatch,
-        languages: List<Language>? = this.languages,
+        languages: List<IsoLanguage>? = this.languages,
         limit: Int? = this.limit,
         requestDebounce: Int? = this.requestDebounce,
         origin: Point? = this.origin,
@@ -233,7 +234,7 @@ public class CategorySearchOptions @JvmOverloads public constructor(
         private var boundingBox: BoundingBox? = null
         private var countries: List<Country>? = null
         private var fuzzyMatch: Boolean? = null
-        private var languages: List<Language>? = defaultSearchOptionsLanguage()
+        private var languages: List<IsoLanguage>? = defaultSearchOptionsLanguage()
         private var limit: Int? = null
         private var requestDebounce: Int? = null
         private var origin: Point? = null
@@ -292,7 +293,7 @@ public class CategorySearchOptions @JvmOverloads public constructor(
          *
          * Note: Geocoding API supports a few languages, Single Box Search – only one.
          */
-        public fun languages(vararg languages: Language): Builder = apply { this.languages = languages.toList() }
+        public fun languages(vararg languages: IsoLanguage): Builder = apply { this.languages = languages.toList() }
 
         /**
          * Specify the user’s language. This parameter controls the language of the text supplied in responses, and also affects result scoring, with results matching the user’s query in the requested language being preferred over results that match in another language. For example, an autocomplete query for things that start with Frank might return Frankfurt as the first result with an English (en) language parameter, but Frankreich (“France”) with a German (de) language parameter.
@@ -300,7 +301,7 @@ public class CategorySearchOptions @JvmOverloads public constructor(
          *
          * Note: Geocoding API supports a few languages, Single Box Search – only one.
          */
-        public fun languages(languages: List<Language>): Builder = apply { this.languages = languages }
+        public fun languages(languages: List<IsoLanguage>): Builder = apply { this.languages = languages }
 
         /**
          * Specify the maximum number of results to return. The maximum supported is 10.