+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/ReverseGeoOptions.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import com.mapbox.geojson.Point
 import com.mapbox.search.base.core.CoreReverseGeoOptions
 import com.mapbox.search.base.core.CoreReverseMode
+import com.mapbox.search.common.IsoLanguage
 import kotlinx.parcelize.Parcelize
 
 /**
@@ -45,7 +46,7 @@ public class ReverseGeoOptions @JvmOverloads public constructor(
      * Specify the user’s language. This parameter controls the language of the text supplied in responses, and also affects result scoring, with results matching the user’s query in the requested language being preferred over results that match in another language. For example, an autocomplete query for things that start with Frank might return Frankfurt as the first result with an English (en) language parameter, but Frankreich (“France”) with a German (de) language parameter.
      * If language is not set explicitly, then language from default system locale will be used.
      */
-    public val languages: List<Language>? = defaultSearchOptionsLanguage(),
+    public val languages: List<IsoLanguage>? = defaultSearchOptionsLanguage(),
 
     /**
      * Specify the maximum number of results to return. The default is 1 and the maximum supported is 5.
@@ -77,7 +78,7 @@ public class ReverseGeoOptions @JvmOverloads public constructor(
     public fun copy(
         center: Point = this.center,
         countries: List<Country>? = this.countries,
-        languages: List<Language>? = this.languages,
+        languages: List<IsoLanguage>? = this.languages,
         limit: Int? = this.limit,
         reverseMode: ReverseMode? = this.reverseMode,
         types: List<QueryType>? = this.types,
@@ -152,7 +153,7 @@ public class ReverseGeoOptions @JvmOverloads public constructor(
     public class Builder(private val center: Point) {
 
         private var countries: List<Country>? = null
-        private var languages: List<Language>? = defaultSearchOptionsLanguage()
+        private var languages: List<IsoLanguage>? = defaultSearchOptionsLanguage()
         private var limit: Int? = null
         private var reverseMode: ReverseMode? = null
         private var types: List<QueryType>? = null
@@ -182,12 +183,12 @@ public class ReverseGeoOptions @JvmOverloads public constructor(
         /**
          * Specify the user’s language. This parameter controls the language of the text supplied in responses, and also affects result scoring, with results matching the user’s query in the requested language being preferred over results that match in another language. For example, an autocomplete query for things that start with Frank might return Frankfurt as the first result with an English (en) language parameter, but Frankreich (“France”) with a German (de) language parameter.
          */
-        public fun languages(vararg languages: Language): Builder = apply { this.languages = languages.toList() }
+        public fun languages(vararg languages: IsoLanguage): Builder = apply { this.languages = languages.toList() }
 
         /**
          * Specify the user’s language. This parameter controls the language of the text supplied in responses, and also affects result scoring, with results matching the user’s query in the requested language being preferred over results that match in another language. For example, an autocomplete query for things that start with Frank might return Frankfurt as the first result with an English (en) language parameter, but Frankreich (“France”) with a German (de) language parameter.
          */
-        public fun languages(languages: List<Language>): Builder = apply { this.languages = languages }
+        public fun languages(languages: List<IsoLanguage>): Builder = apply { this.languages = languages }
 
         /**
          * Specify the maximum number of results to return. The default is 1 and the maximum supported is 5.