+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/SearchOptionsTest.kt
@@ -3,6 +3,7 @@ package com.mapbox.search
 import com.mapbox.geojson.BoundingBox
 import com.mapbox.geojson.Point
 import com.mapbox.search.base.utils.extension.mapToCore
+import com.mapbox.search.common.IsoLanguage
 import com.mapbox.test.dsl.TestCase
 import org.junit.After
 import org.junit.Before
@@ -36,7 +37,7 @@ internal class SearchOptionsTest {
                     boundingBox = null,
                     countries = null,
                     fuzzyMatch = null,
-                    languages = listOf(Language(TEST_LOCALE.language)),
+                    languages = listOf(IsoLanguage(TEST_LOCALE.language)),
                     limit = null,
                     types = null,
                     requestDebounce = null,
@@ -64,7 +65,7 @@ internal class SearchOptionsTest {
                     .countries(Country.UNITED_KINGDOM, Country.BELARUS)
                     .requestDebounce(300)
                     .fuzzyMatch(true)
-                    .languages(Language.ENGLISH, Language("by"))
+                    .languages(IsoLanguage.ENGLISH, IsoLanguage("by"))
                     .limit(100)
                     .types(QueryType.ADDRESS, QueryType.COUNTRY, QueryType.DISTRICT)
                     .origin(TEST_ORIGIN_POINT)
@@ -80,7 +81,7 @@ internal class SearchOptionsTest {
                     boundingBox = TEST_BOUNDING_BOX,
                     countries = listOf(Country.UNITED_KINGDOM, Country.BELARUS),
                     fuzzyMatch = true,
-                    languages = listOf(Language.ENGLISH, Language("by")),
+                    languages = listOf(IsoLanguage.ENGLISH, IsoLanguage("by")),
                     limit = 100,
                     types = listOf(QueryType.ADDRESS, QueryType.COUNTRY, QueryType.DISTRICT),
                     requestDebounce = 300,
@@ -108,7 +109,7 @@ internal class SearchOptionsTest {
                     .countries(listOf(Country.UNITED_KINGDOM, Country.BELARUS))
                     .requestDebounce(300)
                     .fuzzyMatch(true)
-                    .languages(listOf(Language.ENGLISH, Language("by")))
+                    .languages(listOf(IsoLanguage.ENGLISH, IsoLanguage("by")))
                     .limit(100)
                     .types(listOf(QueryType.ADDRESS, QueryType.COUNTRY, QueryType.DISTRICT))
                     .origin(TEST_ORIGIN_POINT)
@@ -124,7 +125,7 @@ internal class SearchOptionsTest {
                     boundingBox = TEST_BOUNDING_BOX,
                     countries = listOf(Country.UNITED_KINGDOM, Country.BELARUS),
                     fuzzyMatch = true,
-                    languages = listOf(Language.ENGLISH, Language("by")),
+                    languages = listOf(IsoLanguage.ENGLISH, IsoLanguage("by")),
                     limit = 100,
                     types = listOf(QueryType.ADDRESS, QueryType.COUNTRY, QueryType.DISTRICT),
                     requestDebounce = 300,
@@ -167,7 +168,7 @@ internal class SearchOptionsTest {
                     .countries(listOf(Country.UNITED_KINGDOM, Country.BELARUS))
                     .requestDebounce(300)
                     .fuzzyMatch(true)
-                    .languages(listOf(Language.ENGLISH, Language("by")))
+                    .languages(listOf(IsoLanguage.ENGLISH, IsoLanguage("by")))
                     .limit(100)
                     .types(listOf(QueryType.ADDRESS, QueryType.COUNTRY, QueryType.DISTRICT))
                     .origin(TEST_ORIGIN_POINT)
@@ -188,7 +189,7 @@ internal class SearchOptionsTest {
                     bbox = TEST_BOUNDING_BOX.mapToCore(),
                     countries = listOf(Country.UNITED_KINGDOM, Country.BELARUS).map { it.code },
                     fuzzyMatch = true,
-                    language = listOf(Language.ENGLISH, Language("by")).map { it.code },
+                    language = listOf(IsoLanguage.ENGLISH, IsoLanguage("by")).map { it.code },
                     limit = 100,
                     types = listOf(QueryType.ADDRESS, QueryType.COUNTRY, QueryType.DISTRICT).map { it.mapToCore() },
                     ignoreUR = true,
@@ -214,7 +215,7 @@ internal class SearchOptionsTest {
                     boundingBox = TEST_BOUNDING_BOX,
                     countries = listOf(Country.UNITED_KINGDOM, Country.BELARUS),
                     fuzzyMatch = true,
-                    languages = listOf(Language.ENGLISH, Language("by")),
+                    languages = listOf(IsoLanguage.ENGLISH, IsoLanguage("by")),
                     limit = 100,
                     types = listOf(QueryType.ADDRESS, QueryType.COUNTRY, QueryType.DISTRICT),
                     requestDebounce = 300,